requirements.txt
pygame
pygame-menu
brainflow
torch
pyzmq
matplotlib
mne
